Class Ast.ast_mapper

Template for defining functions that map over the AST while preserving source locations. This is used by the functions which traverse the de bruijn AST and perform the shift, substitute and normalize operations on de bruijn ASTs.

This inherits from the map class which is automatically generated by the Visitors library. Note that we don't need to impement the visit_App, visit_Type, visit_Kind and visit_Ascription methods since the Visitors package automatically handles all that boilerplate for us.

We only need to override methods that visit AST nodes which contain binders since that is where we need to do something that differs from the default visiting strategy.

`env` is a parameter that is threaded through all the mapping operations. We use this to carry information as we traverse deeper into our AST.

constraint < build_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'h raw_expr Location.located -> Location.source_loc -> 'g expr -> 'g expr; visit_'a : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd -> 'h raw_expr Location.located; visit_App : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Ascription : 'd Location.located -> 'g expr -> 'g expr -> 'g raw_expr; visit_Exists : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Exists_elim : 'd Location.located -> 'g expr -> string -> string -> 'g expr -> 'g raw_expr; visit_Exists_elim_body : 'd Location.located -> string -> string -> 'g expr -> 'g expr; visit_Exists_pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Fst : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Fun : 'd Location.located -> string -> 'g expr option -> 'g expr -> 'g raw_expr; visit_Inl : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Inr : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Kind : 'd Location.located -> 'g raw_expr; visit_Let : 'd Location.located -> 'g abstraction -> 'g expr option -> 'g raw_expr; visit_Match : 'd Location.located -> 'g expr -> 'g match_binding -> 'g match_binding -> 'g raw_expr; visit_Pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Pi : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Sigma : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Snd : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Sum : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Type : 'd Location.located -> 'g raw_expr; visit_Var : 'd Location.located -> 'g -> 'g raw_expr; visit_abstraction : 'd Location.located -> 'g abstraction -> 'g abstraction; visit_expr : 'd Location.located -> 'g expr -> 'g expr; visit_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd Location.located -> 'g expr -> 'g expr; visit_match_binding : 'd Location.located -> 'g match_binding -> 'g match_binding; visit_pair : 'd Location.located -> 'g pair -> 'g pair; visit_raw_expr : 'd Location.located -> 'g raw_expr -> 'g raw_expr; .. > = < build_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'h raw_expr Location.located -> Location.source_loc -> 'g expr -> 'g expr; visit_'a : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd -> 'h raw_expr Location.located; visit_App : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Ascription : 'd Location.located -> 'g expr -> 'g expr -> 'g raw_expr; visit_Exists : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Exists_elim : 'd Location.located -> 'g expr -> string -> string -> 'g expr -> 'g raw_expr; visit_Exists_elim_body : 'd Location.located -> string -> string -> 'g expr -> 'g expr; visit_Exists_pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Fst : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Fun : 'd Location.located -> string -> 'g expr option -> 'g expr -> 'g raw_expr; visit_Inl : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Inr : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Kind : 'd Location.located -> 'g raw_expr; visit_Let : 'd Location.located -> 'g abstraction -> 'g expr option -> 'g raw_expr; visit_Match : 'd Location.located -> 'g expr -> 'g match_binding -> 'g match_binding -> 'g raw_expr; visit_Pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Pi : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Sigma : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Snd : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Sum : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Type : 'd Location.located -> 'g raw_expr; visit_Var : 'd Location.located -> 'g -> 'g raw_expr; visit_abstraction : 'd Location.located -> 'g abstraction -> 'g abstraction; visit_expr : 'd Location.located -> 'g expr -> 'g expr; visit_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd Location.located -> 'g expr -> 'g expr; visit_match_binding : 'd Location.located -> 'g match_binding -> 'g match_binding; visit_pair : 'd Location.located -> 'g pair -> 'g pair; visit_raw_expr : 'd Location.located -> 'g raw_expr -> 'g raw_expr; .. >

inherit map

method visit_expr : < build_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'h raw_expr Location.located -> Location.source_loc -> 'g expr -> 'g expr; visit_'a : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd -> 'h raw_expr Location.located; visit_App : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Ascription : 'd Location.located -> 'g expr -> 'g expr -> 'g raw_expr; visit_Exists : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Exists_elim : 'd Location.located -> 'g expr -> string -> string -> 'g expr -> 'g raw_expr; visit_Exists_elim_body : 'd Location.located -> string -> string -> 'g expr -> 'g expr; visit_Exists_pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Fst : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Fun : 'd Location.located -> string -> 'g expr option -> 'g expr -> 'g raw_expr; visit_Inl : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Inr : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Kind : 'd Location.located -> 'g raw_expr; visit_Let : 'd Location.located -> 'g abstraction -> 'g expr option -> 'g raw_expr; visit_Match : 'd Location.located -> 'g expr -> 'g match_binding -> 'g match_binding -> 'g raw_expr; visit_Pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Pi : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Sigma : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Snd : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Sum : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Type : 'd Location.located -> 'g raw_expr; visit_Var : 'd Location.located -> 'g -> 'g raw_expr; visit_abstraction : 'd Location.located -> 'g abstraction -> 'g abstraction; visit_expr : 'd Location.located -> 'g expr -> 'g expr; visit_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd Location.located -> 'g expr -> 'g expr; visit_match_binding : 'd Location.located -> 'g match_binding -> 'g match_binding; visit_pair : 'd Location.located -> 'g pair -> 'g pair; visit_raw_expr : 'd Location.located -> 'g raw_expr -> 'g raw_expr; .. > -> 'd Location.located -> 'g expr -> 'g expr

method visit_Exists_elim_body : < build_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'h raw_expr Location.located -> Location.source_loc -> 'g expr -> 'g expr; visit_'a : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd -> 'h raw_expr Location.located; visit_App : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Ascription : 'd Location.located -> 'g expr -> 'g expr -> 'g raw_expr; visit_Exists : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Exists_elim : 'd Location.located -> 'g expr -> string -> string -> 'g expr -> 'g raw_expr; visit_Exists_elim_body : 'd Location.located -> string -> string -> 'g expr -> 'g expr; visit_Exists_pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Fst : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Fun : 'd Location.located -> string -> 'g expr option -> 'g expr -> 'g raw_expr; visit_Inl : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Inr : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Kind : 'd Location.located -> 'g raw_expr; visit_Let : 'd Location.located -> 'g abstraction -> 'g expr option -> 'g raw_expr; visit_Match : 'd Location.located -> 'g expr -> 'g match_binding -> 'g match_binding -> 'g raw_expr; visit_Pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Pi : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Sigma : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Snd : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Sum : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Type : 'd Location.located -> 'g raw_expr; visit_Var : 'd Location.located -> 'g -> 'g raw_expr; visit_abstraction : 'd Location.located -> 'g abstraction -> 'g abstraction; visit_expr : 'd Location.located -> 'g expr -> 'g expr; visit_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd Location.located -> 'g expr -> 'g expr; visit_match_binding : 'd Location.located -> 'g match_binding -> 'g match_binding; visit_pair : 'd Location.located -> 'g pair -> 'g pair; visit_raw_expr : 'd Location.located -> 'g raw_expr -> 'g raw_expr; .. > -> 'd Location.located -> string -> string -> 'g expr -> 'g expr

method visit_Exists_elim : < build_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'h raw_expr Location.located -> Location.source_loc -> 'g expr -> 'g expr; visit_'a : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd -> 'h raw_expr Location.located; visit_App : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Ascription : 'd Location.located -> 'g expr -> 'g expr -> 'g raw_expr; visit_Exists : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Exists_elim : 'd Location.located -> 'g expr -> string -> string -> 'g expr -> 'g raw_expr; visit_Exists_elim_body : 'd Location.located -> string -> string -> 'g expr -> 'g expr; visit_Exists_pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Fst : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Fun : 'd Location.located -> string -> 'g expr option -> 'g expr -> 'g raw_expr; visit_Inl : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Inr : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Kind : 'd Location.located -> 'g raw_expr; visit_Let : 'd Location.located -> 'g abstraction -> 'g expr option -> 'g raw_expr; visit_Match : 'd Location.located -> 'g expr -> 'g match_binding -> 'g match_binding -> 'g raw_expr; visit_Pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Pi : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Sigma : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Snd : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Sum : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Type : 'd Location.located -> 'g raw_expr; visit_Var : 'd Location.located -> 'g -> 'g raw_expr; visit_abstraction : 'd Location.located -> 'g abstraction -> 'g abstraction; visit_expr : 'd Location.located -> 'g expr -> 'g expr; visit_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd Location.located -> 'g expr -> 'g expr; visit_match_binding : 'd Location.located -> 'g match_binding -> 'g match_binding; visit_pair : 'd Location.located -> 'g pair -> 'g pair; visit_raw_expr : 'd Location.located -> 'g raw_expr -> 'g raw_expr; .. > -> 'd Location.located -> 'g expr -> string -> string -> 'g expr -> 'g raw_expr

method visit_'a : < build_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'h raw_expr Location.located -> Location.source_loc -> 'g expr -> 'g expr; visit_'a : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd -> 'h raw_expr Location.located; visit_App : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Ascription : 'd Location.located -> 'g expr -> 'g expr -> 'g raw_expr; visit_Exists : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Exists_elim : 'd Location.located -> 'g expr -> string -> string -> 'g expr -> 'g raw_expr; visit_Exists_elim_body : 'd Location.located -> string -> string -> 'g expr -> 'g expr; visit_Exists_pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Fst : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Fun : 'd Location.located -> string -> 'g expr option -> 'g expr -> 'g raw_expr; visit_Inl : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Inr : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Kind : 'd Location.located -> 'g raw_expr; visit_Let : 'd Location.located -> 'g abstraction -> 'g expr option -> 'g raw_expr; visit_Match : 'd Location.located -> 'g expr -> 'g match_binding -> 'g match_binding -> 'g raw_expr; visit_Pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Pi : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Sigma : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Snd : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Sum : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Type : 'd Location.located -> 'g raw_expr; visit_Var : 'd Location.located -> 'g -> 'g raw_expr; visit_abstraction : 'd Location.located -> 'g abstraction -> 'g abstraction; visit_expr : 'd Location.located -> 'g expr -> 'g expr; visit_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd Location.located -> 'g expr -> 'g expr; visit_match_binding : 'd Location.located -> 'g match_binding -> 'g match_binding; visit_pair : 'd Location.located -> 'g pair -> 'g pair; visit_raw_expr : 'd Location.located -> 'g raw_expr -> 'g raw_expr; .. > -> ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd -> 'h raw_expr Location.located

method build_located : < build_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'h raw_expr Location.located -> Location.source_loc -> 'g expr -> 'g expr; visit_'a : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd -> 'h raw_expr Location.located; visit_App : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Ascription : 'd Location.located -> 'g expr -> 'g expr -> 'g raw_expr; visit_Exists : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Exists_elim : 'd Location.located -> 'g expr -> string -> string -> 'g expr -> 'g raw_expr; visit_Exists_elim_body : 'd Location.located -> string -> string -> 'g expr -> 'g expr; visit_Exists_pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Fst : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Fun : 'd Location.located -> string -> 'g expr option -> 'g expr -> 'g raw_expr; visit_Inl : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Inr : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Kind : 'd Location.located -> 'g raw_expr; visit_Let : 'd Location.located -> 'g abstraction -> 'g expr option -> 'g raw_expr; visit_Match : 'd Location.located -> 'g expr -> 'g match_binding -> 'g match_binding -> 'g raw_expr; visit_Pair : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Pi : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Sigma : 'd Location.located -> 'g abstraction -> 'g raw_expr; visit_Snd : 'd Location.located -> 'g expr -> 'g raw_expr; visit_Sum : 'd Location.located -> 'g pair -> 'g raw_expr; visit_Type : 'd Location.located -> 'g raw_expr; visit_Var : 'd Location.located -> 'g -> 'g raw_expr; visit_abstraction : 'd Location.located -> 'g abstraction -> 'g abstraction; visit_expr : 'd Location.located -> 'g expr -> 'g expr; visit_located : ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'd Location.located -> 'g expr -> 'g expr; visit_match_binding : 'd Location.located -> 'g match_binding -> 'g match_binding; visit_pair : 'd Location.located -> 'g pair -> 'g pair; visit_raw_expr : 'd Location.located -> 'g raw_expr -> 'g raw_expr; .. > -> ('d Location.located -> 'g raw_expr -> 'g raw_expr) -> 'h raw_expr Location.located -> Location.source_loc -> 'g expr -> 'g expr