Module Common.Location

type source_loc = Stdlib.Lexing.position * Stdlib.Lexing.position

This datatype is used to wrap our ASTs with source location information.

type 'a located = {

data : 'a;

source_loc : source_loc;

}

val pp_located : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a located -> Ppx_deriving_runtime.unit

val show_located : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a located -> Ppx_deriving_runtime.string

val equal_located : a. ('a -> 'a -> Ppx_deriving_runtime.bool) -> 'a located -> 'a located -> Ppx_deriving_runtime.bool

include sig ... end

class virtual +'a fold : object ... end

val source_loc : 'a located -> source_loc

val data : 'a located -> 'a

module Fields_of_located : sig ... end

val locate : ?⁠source_loc:source_loc -> 'a -> 'a located

locate ?source_loc data wraps data in a located record using source_loc.

If source_loc is not provided, a dummy location is used.

val update_data : 'a located -> ('a -> 'b) -> 'b located

val set_data : 'a located -> 'b -> 'b located

set_data located new_data sets the data in located to new_data, ignoring the old data.