Module Common__Ast

type 'a expr = 'a raw_expr Common.Location.located

Expressions are raw expressions together with their source locations.

and 'a abstraction = {

var_name : string;

expr : 'a expr;

body : 'a expr;

}

This is used to represent the Pi, Sigma and Existential types since they all have similar structure:

Pi (var_name : expr), body

Sigma (var_name : expr), body

exists (var_name : expr), body

Note that var_name is bound in body but not expr.

and 'a match_binding = {

match_var : string;

match_body : 'a expr;

}

Used to represent the 2 clauses of match expressions, ie:

inl match_var => match_body

inr match_var => match_body

and 'a pair = {

left : 'a expr;

right : 'a expr;

}

Used to represent

Function application, App: (left right)

Pair constructor for Sigma types, Pair: (left, right)

Type constructor for Sum type: left + right

Pair constuctor for Existential types, Exists_pair: {left, right}

and 'a raw_expr =

| Type

| Kind

| Var of 'a
Variables 'a is string for the parser's AST and int for the internal AST using de bruijn indices.

| Pi of 'a abstraction
Pi (var_name : expr), body

| Fun of {

input_var : string;

input_type : 'a expr option;

body : 'a expr;

}
Represents:

fun input_var => body

fun (input_var : input_type) => body

| App of 'a pair
Function application: (left right)

| Sigma of 'a abstraction
Sigma type: Sigma (var_name : expr), body

| Pair of 'a pair
Pair constructor: (left, right)

| Fst of 'a expr
Pair eliminator: fst expr

| Snd of 'a expr
Pair eliminator: snd expr

| Exists_pair of 'a pair
Pair constuctor for Existential types: {left, right}

| Exists of 'a abstraction
Existential type: exists (var_name : expr), body

| Exists_elim of {

expr : 'a expr;

witness_var : string;

witness_cert : string;

body : 'a expr;

}
Existential elimination:

let {witness_var, witness_cert} := expr in body

| Sum of 'a pair
Sum type: left + right

| Inl of 'a expr
Left constructor for sum type: inl expr

| Inr of 'a expr
Right constructor for sum type: inr expr

| Match of {

expr : 'a expr;

inl : 'a match_binding;

inr : 'a match_binding;

}
Sum eliminator:

match expr with

| inl match_var => match_body

| inr match_var => match_body

end

| Ascription of {

expr : 'a expr;

ascribed_type : 'a expr;

}
Type ascriptions: (expr : ascribed_type)

| Let of {

abstraction : 'a abstraction;

ascribed_type : 'a expr option;

}
Local let binding:

let var_name := expr in body

let var_name : ascribed_type := expr in body

val pp_expr : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a expr -> Ppx_deriving_runtime.unit

val show_expr : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a expr -> Ppx_deriving_runtime.string

val pp_abstraction : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a abstraction -> Ppx_deriving_runtime.unit

val show_abstraction : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a abstraction -> Ppx_deriving_runtime.string

val pp_match_binding : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a match_binding -> Ppx_deriving_runtime.unit

val show_match_binding : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a match_binding -> Ppx_deriving_runtime.string

val pp_pair : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a pair -> Ppx_deriving_runtime.unit

val show_pair : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a pair -> Ppx_deriving_runtime.string

val pp_raw_expr : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a raw_expr -> Ppx_deriving_runtime.unit

val show_raw_expr : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a raw_expr -> Ppx_deriving_runtime.string

val equal_expr : a. ('a -> 'a -> Ppx_deriving_runtime.bool) -> 'a expr -> 'a expr -> Ppx_deriving_runtime.bool

val equal_abstraction : a. ('a -> 'a -> Ppx_deriving_runtime.bool) -> 'a abstraction -> 'a abstraction -> Ppx_deriving_runtime.bool

val equal_match_binding : a. ('a -> 'a -> Ppx_deriving_runtime.bool) -> 'a match_binding -> 'a match_binding -> Ppx_deriving_runtime.bool

val equal_pair : a. ('a -> 'a -> Ppx_deriving_runtime.bool) -> 'a pair -> 'a pair -> Ppx_deriving_runtime.bool

val equal_raw_expr : a. ('a -> 'a -> Ppx_deriving_runtime.bool) -> 'a raw_expr -> 'a raw_expr -> Ppx_deriving_runtime.bool

include sig ... end

class virtual +'c map : object ... end

include sig ... end

class virtual +'a fold : object ... end

include sig ... end

class virtual +'b iter : object ... end

val body : 'a abstraction -> 'a expr

val expr : 'a abstraction -> 'a expr

val var_name : 'a abstraction -> string

module Fields_of_abstraction : sig ... end

val match_body : 'a match_binding -> 'a expr

val match_var : 'a match_binding -> string

module Fields_of_match_binding : sig ... end

val right : 'a pair -> 'a expr

val left : 'a pair -> 'a expr

module Fields_of_pair : sig ... end

val located_kind : 'a raw_expr Common.Location.located

class virtual +'c ast_mapper : object ... end

Template for defining functions that map over the AST while preserving source locations. This is used by the functions which traverse the de bruijn AST and perform the shift, substitute and normalize operations on de bruijn ASTs.

class virtual +'a ast_folder : object ... end

Template for defining folds over the AST that ignore source locations.

val shift : int -> int expr -> int expr

shift shift_by expr is the shift operation for de bruijn ASTs.

Note that whenever we go under a binder, the cutoff index must be incremented by 1. If we go under 2 binders, as in the case of existential elimination expressions,

let {x, y} := E in E'

then we must increase the cutoff by 2.

val do_if_index_present : int -> (int -> unit) -> int expr -> unit

do_if_index_present index f traverses a de bruijn AST in search for a variable of index index. If any such variables are found, f is called with index as the input.

Note that index is incremented whenever we go under a binder.

type 'a list_of_exprs = 'a expr list

List of expressions

val pp_list_of_exprs : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a list_of_exprs -> Ppx_deriving_runtime.unit

val show_list_of_exprs : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a list_of_exprs -> Ppx_deriving_runtime.string

type 'a stmt = 'a raw_stmt Common.Location.located

As with expressions, statements come with a raw statement, raw_stmt wrapped with a source location.

and 'a raw_stmt =

| Def of {

var_name : string;

binding : 'a expr;

ascribed_type : 'a expr option;

}
Def statement:

def var_name := binding

def var_name : ascribed_type := binding

| Axiom of {

var_name : string;

var_type : 'a expr;

}
Axiom statement:

axiom var_name : var_type

| Check of 'a expr
Check statement:

check expr

| Eval of 'a expr
Eval statement:

eval expr

val pp_stmt : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a stmt -> Ppx_deriving_runtime.unit

val show_stmt : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a stmt -> Ppx_deriving_runtime.string

val pp_raw_stmt : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a raw_stmt -> Ppx_deriving_runtime.unit

val show_raw_stmt : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a raw_stmt -> Ppx_deriving_runtime.string

type 'a list_of_stmts = 'a stmt list

List of statements

val pp_list_of_stmts : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a list_of_stmts -> Ppx_deriving_runtime.unit

val show_list_of_stmts : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a list_of_stmts -> Ppx_deriving_runtime.string