Module Lambdacplus.Toplevel

type contexts = {

naming_ctx : Kernel.Context.t;
This is used to parse the string to the parser's AST and then convert it to the de bruijn one, where all variables are identified by de bruijn indices.

eval_ctx : Kernel.Context.t;
This is used for typechecking and evluating expressions in the de bruijn AST. All types and bindings stored here have indices for variable names, rather than strings.

}

The current state of the naming and evaluation contexts.

val initial_contexts : contexts

Initial (ie empty) naming and evaluation contexts.

val parse_and_eval : string -> contexts -> (contexts * string, string * Error_reporting.err_loc) Stdlib.result

parse_and_eval str {naming_ctx; eval_ctx} parses and evalutes a string.

Note that we actually use 2 contexts to parse and evaluate user inputs:

naming_ctx is first used to parse the input string into the de bruijn AST.

eval_ctx is then used for typechecking and evaluating.

The output is a result record (which is like an Either monad) that is either

Ok {naming_ctx, eval_ctx}, result where

naming_ctx and eval_ctx are the new naming and evaluation contexts respectively

result is the string of the expression resulting from evaluating str

Error err_str, err_loc where

err_str is a string representation of the error that occurred

err_loc is a Error_reporting.err_loc, ie a record containing the location of the error